pyproject.toml
[project]
name = "helm-release-mcp"
version = "0.1.0"
description = "MCP server for automating Dify EE Helm chart releases across GitHub repositories"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "pydantic-settings>=2.0.0",
    "gitpython>=3.1.0",
    "pygithub>=2.0.0",
    "ruamel.yaml>=0.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "pytest-watch>=4.2.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
helm-release-mcp = "helm_release_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/helm_release_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501", "SIM105"]  # SIM105: contextlib.suppress doesn't work when capturing result

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
